Split `HostUtil` out of `UriUtil`.
