+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Breaking changes:
 * `net-util`:
   * Major rework of `IncomingRequest`, so it no longer has to be constructed
     from a low-level Node request object.
-  * Renamed `Uris` to `UriUtil`.
+  * Split `Uris` into two classes, `UriUtil` and `HostUtil`.
   * Pulled `uriPathStringFrom()` from `TreePathKey` into `UriUtil`, renaming it
     to `pathStringFrom()` and tweaking its contract.
   * Replaced `DispatchInfo.{base,extra}String` properties with a single combined
